new speed up attempt

+++ shap_config.py
@@ -193,7 +193,7 @@ class shap_conf():
         for ii in range(num_struc):
             indx = np.array(np.where(self.segmentation == ii)).transpose()
             struc_indx.append(indx)
-        self.struc_indx = np.array(struc_indx).astype(int)
+        self.struc_indx = np.array(struc_indx)
         
         
     def mask_dom(self,zs):
